Create executable to start component playground via webpack dev server #114
