+++ component-playground/public/component-playground.html
@@ -1,0 +1,12 @@
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>React Component Playground</title>
+  <!-- Require it using style-loader, inside entry.js -->
+  <link rel="stylesheet" href="global-reset.css">
+</head>
+<body>
+  <div id="component-playground"></div>
+  <script src="/assets/bundle.js"></script>
+</body>
+</html>
