+++ component-playground/public/global-reset.css
@@ -1,0 +1,5 @@
+html, body {
+  margin: 0;
+  padding: 0;
+  height: 100%;
+}
