+++ package.json
@@ -35,7 +35,8 @@
     "sinon": "^1.12.2",
     "sinon-chai": "^2.6.0",
     "style-loader": "^0.8.3",
-    "webpack": "^1.5.3"
+    "webpack": "^1.5.3",
+    "webpack-dev-server": "^1.7.0"
   },
   "main": "build/cosmos.commonjs.js",
   "scripts": {
